+++ ttsd-service/src/main/java/com/tuotiansudai/repository/model/TitleModel.java
@@ -6,17 +6,17 @@ import java.math.BigInteger;
  * Created by tuotian on 15/8/17.
  */
 public class TitleModel {
-    private BigInteger id;
+    private long id;
     /***标题类型base:基础标题，new:新增标题***/
     private String type;
     /***标题名称***/
     private String title;
 
-    public BigInteger getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(long id) {
         this.id = id;
     }
